Changed Format
Downloaded the actual GitGud repository so I could open the whole game folder instead of having to edit the game in the html file, this should make it easier for everyone as we don't have to dig through Twee.

I will need to make a video tutorial on how to get set up so that all contributors are on the same page and know how to edit stuff, and also do a guided tour of the code.